FILE: moves dtos to top level of lib
