+++ libs/menu-matriarch/shared/data-access-dtos/src/lib/_schema.interface.ts
