+++ libs/menu-matriarch/shared/data-access-dtos/src/lib/dish-dto.ts
