+++ libs/menu-matriarch/shared/data-access-dtos/src/lib/mappers/map-ingredient-type-dto-to-ingredient-type.ts
@@ -1,5 +1,5 @@
 import { Ingredient, IngredientType } from '@atocha/menu-matriarch/shared/util';
-import { IngredientTypeDto } from '../dtos/ingredient-type-dto';
+import { IngredientTypeDto } from '../ingredient-type-dto';
 
 export function mapIngredientTypeDtoToIngredientType(
   ingredientTypeDto: IngredientTypeDto,
