+++ libs/menu-matriarch/shared/data-access-dtos/src/lib/mappers/map-meal-dto-to-meal.ts
@@ -1,5 +1,5 @@
 import { Dish, Tag, Meal } from '@atocha/menu-matriarch/shared/util';
-import { MealDto } from '../dtos/meal-dto';
+import { MealDto } from '../meal-dto';
 
 export function mapMealDtoToMeal({
   mealDto,
